separate navigation items into components
