+++ pages/index.vue
@@ -11,7 +11,7 @@
 <script>
   import ModalViewGene from '~/components/ModalView/ModalViewGene.vue';
   import ModalViewCompare from '~/components/ModalView/ModalViewCompare.vue';
-  import LocalNavigation from '~/components/search/LocalNavigation.vue';
+  import LocalNavigation from '~/components/search/LocalNavigation/LocalNavigation.vue';
   import FilterTab from '~/components/search/FilterTab.vue';
   import filters from '~/static/filters.json';
   import ModalViewAlert from '~/components/ModalView/ModalViewAlert.vue';
